Use sized types in 64-bit int/uint serialize
On x86_64  linux, for example, uint64_t is defined as 'long', not 'long
long', which has a different type according to the c++ mangling (despite
being the same size and everything) so fails to link

+++ game/state/gamestate_serialize.cpp
@@ -1,3 +1,4 @@
+#include "game/state/gamestate.h"
 #include "framework/framework.h"
 #include "framework/image.h"
 #include "framework/serialization/serialize.h"
@@ -12,7 +13,6 @@
 #include "game/state/city/vehicle.h"
 #include "game/state/city/vehiclemission.h"
 #include "game/state/city/vequipment.h"
-#include "game/state/gamestate.h"
 #include "game/state/rules/scenery_tile_type.h"
 #include "game/state/rules/vequipment_type.h"
 #include "library/voxel.h"
@@ -68,14 +68,14 @@ void serializeIn(const GameState *state, sp<SerializationNode> node, int &val)
 	val = node->getValueInt();
 }
 
-void serializeIn(const GameState *state, sp<SerializationNode> node, unsigned long long &val)
+void serializeIn(const GameState *state, sp<SerializationNode> node, uint64_t &val)
 {
 	if (!node)
 		return;
 	val = node->getValueUInt64();
 }
 
-void serializeIn(const GameState *state, sp<SerializationNode> node, long long &val)
+void serializeIn(const GameState *state, sp<SerializationNode> node, int64_t &val)
 {
 	if (!node)
 		return;
@@ -878,12 +878,9 @@ void serializeOut(sp<SerializationNode> node, const float &val) { node->setValue
 
 void serializeOut(sp<SerializationNode> node, const int &val) { node->setValueInt(val); }
 
-void serializeOut(sp<SerializationNode> node, const unsigned long long &val)
-{
-	node->setValueUInt64(val);
-}
+void serializeOut(sp<SerializationNode> node, const uint64_t &val) { node->setValueUInt64(val); }
 
-void serializeOut(sp<SerializationNode> node, const long long &val) { node->setValueInt64(val); }
+void serializeOut(sp<SerializationNode> node, const int64_t &val) { node->setValueInt64(val); }
 
 void serializeOut(sp<SerializationNode> node, const bool &val) { node->setValueBool(val); }
